requirements.txt
apache-airflow>=3.0.6
requests>=2.31.0
dbt-postgres>=1.5.0
psycopg2-binary>=2.9.0
